Move AppInstallMonitorLog
